OPENEUROPA-1192: Replaced validator code with event.
