+++ composer.json
@@ -56,7 +56,8 @@
         "patches": {
           "drupal/cas": {
             "Add support for http CAS servers": "https://www.drupal.org/files/issues/2018-09-03/2996924-4.patch",
-            "Provide custom mail handlers": "https://www.drupal.org/files/issues/2018-09-04/2996955-2.patch"
+            "Provide custom mail handlers": "https://www.drupal.org/files/issues/2018-09-04/2996955-2.patch",
+            "Provide after validation event": "https://www.drupal.org/files/issues/2018-09-05/2997099-4.patch"
           }
         }
     },
